Use MUnit in PipelineSuite
